Add parallel strategies based on the Eval Monad
Now there are four strategies based on the eval monad (there was only
one before). One strategy splits the evaluation at every choice
(splitAll). At every choice the strategy creates a new spark, which
can be evaluated in parallel. The other strategies are used to reduce
the number of sparks that are created:
- splitLimitDepth n splits the evaluation for a depth of n
- splitAlternating n splits at every n-th choice
- splitPower splits at every node with a distance of 2^i from the root